Refactored user module: integrated Express Router and controller pattern
